backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
pytest
pytest-django
pytest-cov
factory-boy

#requirements.txt